pyproject.toml
# PEP 621 standard metadata — works directly with uv (no Poetry required).
#
# Dependency groups (PEP 735) — never published, local development only:
#   uv sync                   → production deps only
#   uv sync --group test      → + pytest, factory-boy, httpx
#   uv sync --group lint      → + black, flake8, isort, mypy
#   uv sync --all-groups      → everything (local dev)

[project]
name = "vm-lifecycle-api"
version = "0.1.0"
description = "OpenStack VM Lifecycle Management API"
readme = "README.md"
requires-python = ">=3.11"

# Production dependencies only — installed in the runtime Docker image.
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "openstacksdk>=3.3.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    "httpx>=0.27.0",
    "redis>=5.0",
    "arq>=0.26.0",
]

# ── Dependency groups (PEP 735) ───────────────────────────────────────────────
# These are NOT published as PyPI extras. uv installs them only when
# explicitly requested with --group / --all-groups.

[dependency-groups]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",          # AsyncClient / TestClient
    "factory-boy>=3.3.0",     # fixture factories (tests/mocks/factories.py)
    "fakeredis>=2.26.0",      # in-process Redis for RedisTaskStore tests
]
lint = [
    "black>=24.0.0",
    "flake8>=7.0.0",
    "flake8-bugbear>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.12.0",
]
dev = [
    "pre-commit>=4.5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=100 -v"
filterwarnings = [
    # fakeredis 2.x uses deprecated redis-py 7.x constructor kwargs internally;
    # nothing we can fix until fakeredis is updated upstream.
    "ignore::DeprecationWarning:fakeredis",
    "ignore:.*retry_on_timeout.*:DeprecationWarning",
]

[tool.coverage.report]
exclude_lines = [
    "@abstractmethod",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
